+++ src/main/java/com/example/whik/entity/WhikApplication.java
@@ -1,4 +1,4 @@
-package com.example.whik;
+package com.example.whik.entity;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
