+++ src/main/java/com/example/whik/entity/Wishlist.java
@@ -1,9 +1,6 @@
-package com.example.whik;
+package com.example.whik.entity;
 
-import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -18,28 +15,29 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class MemberCategory {
+public class Wishlist {
+
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "member_category_id")
 	private Long id;
 
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "member_id")
 	private Member member;
 
-	@Enumerated(EnumType.STRING)
-	private Category category;
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "destination_id")
+	private Destination destination;
 
 	@Builder(access = AccessLevel.PRIVATE)
-	private MemberCategory(Member member, Category category) {
+	private Wishlist(Member member, Destination destination) {
 		this.member = member;
-		this.category = category;
+		this.destination = destination;
 	}
 
-	public static MemberCategory create(Member member, Category category) {
-		return MemberCategory.builder()
+	public static Wishlist create(Member member, Destination destination) {
+		return Wishlist.builder()
 				.member(member)
-				.category(category).build();
+				.destination(destination).build();
 	}
 }
